+++ package.json
@@ -77,10 +77,12 @@
   },
   "dependencies": {
     "@segment/analytics-node": "^1.1.0",
+    "binaryextensions": "^4.18.0",
     "ejs": "^3.1.9",
     "enquirer": "^2.4.0",
     "fs-extra": "^11.1.1",
     "ora": "^5.1.0",
-    "pretty-time": "^1.1.0"
+    "pretty-time": "^1.1.0",
+    "textextensions": "^5.16.0"
   }
 }
